+++ SCME.Types/ProhibitionVoltage/TypeProhibitionVoltage.cs
@@ -27,24 +27,24 @@ namespace SCME.Types.ProhibitionVoltage
         public TypeManagement TypeManagement { get; set; }
 
         [DataMember]
-        public ushort ControlVoltage { get; set; }
+        public double ControlVoltage { get; set; }
         [DataMember]
-        public ushort ControlCurrent { get; set; }
+        public double ControlCurrent { get; set; }
 
         [DataMember]
-        public ushort SwitchedAmperage { get; set; }
+        public double SwitchedAmperage { get; set; }
         [DataMember]
-        public ushort SwitchedVoltage { get; set; }
+        public double SwitchedVoltage { get; set; }
 
         [DataMember]
-        public ushort AuxiliaryVoltagePowerSupply1 { get; set; }
+        public double AuxiliaryVoltagePowerSupply1 { get; set; }
         [DataMember]
-        public ushort AuxiliaryCurrentPowerSupply1 { get; set; }
+        public double AuxiliaryCurrentPowerSupply1 { get; set; }
 
         [DataMember]
-        public ushort AuxiliaryVoltagePowerSupply2 { get; set; }
+        public double AuxiliaryVoltagePowerSupply2 { get; set; }
         [DataMember]
-        public ushort AuxiliaryCurrentPowerSupply2 { get; set; }
+        public double AuxiliaryCurrentPowerSupply2 { get; set; }
 
         [DataMember]
         public double ProhibitionVoltageMinimum { get; set; }
